third_party/blink/web_tests/wpt_internal/route/parsing/url-valid.html

<!DOCTYPE html>
<title>Serialization of url()</title>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-navigation-1/#typedef-route-location">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  test(()=> {
    const canonical = "@navigation (at: url(\"/foo/\")) {\n}";
    test_valid_rule(canonical);
    test_valid_rule("@Navigation  ( At:Url( '/foo/' )){ \n}", canonical);
    test_valid_rule("@Navigation  ( At:Url( /foo/ )){ \n}", canonical);
  }, "<navigation-location-test>");

  test(()=> {
    const canonical = ":active-navigation(at url(\"/foo/\")) { }";
    test_valid_rule(canonical);
    test_valid_rule(":active-navigation( At Url( '/foo/' ) ){}", canonical);
    test_valid_rule(":active-navigation( At Url( /foo/ ) ){}", canonical);
  }, ":active-navigation()");

  test(()=> {
    const canonical = ":active-navigation(with url(\"/foo/\")) { }";
    test_valid_rule(canonical);
    test_valid_rule(":active-navigation(Url( '/foo/' ) ){}", canonical);
    test_valid_rule(":active-navigation(Url( /foo/ ) ){}", canonical);
  }, ":active-navigation() default 'with'");
</script>
